# docs/openapi.yaml
openapi: 3.0.3
info:
  title: LifeFlow API
  description: |
    Personal productivity API for task management, goal tracking, habits,
    focus sessions, lists, notes, and more.

    ## Error Responses
    All error responses use a consistent shape:
    ```json
    { "error": "Human-readable error message" }
    ```

    ## Common Status Codes
    | Code | Meaning |
    |------|---------|
    | 200  | Success |
    | 201  | Created (POST that creates a resource) |
    | 400  | Bad request (validation error) |
    | 404  | Resource not found |
    | 409  | Conflict (constraint violation) |
    | 500  | Internal server error |
  version: 0.7.1
  license:
    name: MIT
servers:
  - url: http://localhost:3456
    description: Local development
...
paths:
...
  /api/filters:
    get:
      tags: [Filters]
      summary: List saved filters
      responses:
        '200':
          description: Array of filters
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/Filter' }
    post:
      tags: [Filters]
      summary: Create a saved filter
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, criteria]
              properties:
                name: { type: string }
                icon: { type: string }
                criteria: { type: object }
      responses:
        '201':
          description: Created filter
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Filter' }
  /api/filters/{id}:
    put:
      tags: [Filters]
      summary: Update a saved filter
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        '200':
          description: Updated filter
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Filter' }
    delete:
      tags: [Filters]
      summary: Delete a saved filter
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        '200':
          description: Deletion confirmation
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok: { type: boolean }
  /api/filters/execute:
    get:
      tags: [Filters]
      summary: Execute a filter and return matching tasks
      parameters:
        - name: status
          in: query
          schema: { type: string }
        - name: priority
          in: query
          schema: { type: integer }
        - name: area_id
          in: query
          schema: { type: integer }
        - name: goal_id
          in: query
          schema: { type: integer }
        - name: tag_id
          in: query
          schema: { type: integer }
        - name: overdue
          in: query
          schema: { type: string, enum: ['true'] }
        - name: has_note
          in: query
          schema: { type: string, enum: ['true'] }
        - name: recurring
          in: query
          schema: { type: string, enum: ['true'] }
      responses:
        '200':
          description: Array of matching tasks
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/Task' }
  /api/filters/smart/{type}:
    get:
      tags: [Filters]
      summary: Get tasks matching a smart filter
      parameters:
        - name: type
          in: path
          required: true
          schema:
            type: string
            enum: [high-priority, due-this-week, no-due-date, completed-today, untagged, blocked]
      responses:
        '200':
          description: Array of matching tasks
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/Task' }
...
components:
...
  schemas:
    Task:
      type: object
      properties:
        id: { type: integer }
        goal_id: { type: integer }
        title: { type: string }
        note: { type: string, nullable: true }
        status: { type: string, enum: [todo, doing, done] }
        priority: { type: integer, minimum: 0, maximum: 3 }
        position: { type: integer }
        due_date: { type: string, format: date, nullable: true }
        due_time: { type: string, nullable: true }
        recurring: { type: string, nullable: true }
        assigned_to: { type: string, nullable: true }
        my_day: { type: integer, enum: [0, 1] }
        created_at: { type: string, format: date-time }
        completed_at: { type: string, format: date-time, nullable: true }
        estimated_minutes: { type: number, nullable: true }
        actual_minutes: { type: number, nullable: true }
        time_block_start: { type: string, nullable: true }
        time_block_end: { type: string, nullable: true }
        list_id: { type: integer, nullable: true }
        reminder_at: { type: string, nullable: true }
        tags:
          type: array
          items: { $ref: '#/components/schemas/Tag' }
        subtasks:
          type: array
          items: { $ref: '#/components/schemas/Subtask' }
        blocked_by:
          type: array
          items:
            type: object
            properties:
              id: { type: integer }
              title: { type: string }
              status: { type: string }
    Subtask:
      type: object
      properties:
        id: { type: integer }
        task_id: { type: integer }
        title: { type: string }
        done: { type: integer, enum: [0, 1] }
        position: { type: integer }
    Tag:
      type: object
      properties:
        id: { type: integer }
        name: { type: string }
        color: { type: string, nullable: true }
...
    Filter:
      type: object
      properties:
        id: { type: integer }
        name: { type: string }
        icon: { type: string, nullable: true }
        criteria: { type: object }
        created_at: { type: string, format: date-time }
...